+++ code/src/cljs/sixsq/nuvla/ui/main/events.cljs
@@ -21,10 +21,11 @@
 
 (reg-event-db
   ::set-device
-  (fn [db [_ device]]
-    (log/info "setting device:" device)
-    (cond-> (assoc db ::spec/device device)
-            (not (#{:mobile :table} device)) (assoc ::spec/sidebar-open? true))))
+  (fn [{:keys [::spec/device] :as db} [_ new-device]]
+    (log/info "setting device:" new-device)
+    (cond-> (assoc db ::spec/device new-device)
+            (not= device new-device) (assoc ::spec/sidebar-open?
+                                            (not (#{:mobile :tablet} new-device))))))
 
 
 (reg-event-db